Added some changes to the export file to make it look better. Removed duplicate mouse listener from FTCauto. Made some corrections to the code for exporting in MenuBars. Added initialization to ProgramInfo constructor.

+++ AnimationTest/src/animationtest/RobotEditorFrame.java
@@ -304,7 +304,7 @@ public class RobotEditorFrame extends javax.swing.JFrame {
         // TODO add your handling code here:
         data.distanceBetweenWheels = Double.parseDouble(wheelDistance.getText());
         data.wheelDiameter = Double.parseDouble(wheelDiameter.getText());
-        data.gearRatio = Integer.parseInt(gearRatio1.getText()) / Integer.parseInt(gearRatio2.getText());
+        data.gearRatio = Integer.parseInt(gearRatio1.getText()) / (double) Integer.parseInt(gearRatio2.getText());
         data.clearMotors();
         if (software1.getText().length() > 0 && hardware1.getText().length() > 0) {
             data.getMotors().add(new ProgramInfo.ItemData(software1.getText(), hardware1.getText(), reversed1.isSelected(), true));
